pyproject.toml
[tool.poetry]
name = "rosalind_solver"
version = "0.0.1"
description = "Solves rosalind problems with custom inputs"
authors = ["TimBerneiser <tim@berneiser.de>"]
license = "MIT"
readme = "README.rst"

[tool.poetry.dependencies]
python = "^3.10"
biopython = "^1.83"
pytest = "^8.2.1"
pylint = "^3.2.2"
flake8 = "^7.0.0"
pandas = "^2.2.2"
numpy = "^1.26.4"
ttkbootstrap = "^1.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
